% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialization of the Composite EM Algorithm.R
\name{em_init}
\alias{em_init}
\title{Initialize the coefficient and the baseline hazard terms}
\usage{
em_init(p, tau)
}
\arguments{
\item{p}{dimension of the coefficient vector}

\item{tau}{list of length S containing tau.s for each stratum}
}
\value{
A list with two components
\describe{
\item{beta}{A list of vector 0s with length of \code{p}, representing initialized coefficients }
\item{lambda}{a list of vectors where each vector has an initial baseline hazard value (1/length of tau.s) } }
}
\description{
Initialize the coefficient and the baseline hazard terms
}
