% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Data_Generation_random_baseline.R
\name{g}
\alias{g}
\title{Positive Stable Distribution Constant}
\usage{
g(a)
}
\arguments{
\item{a}{Numeric. Index parameter \eqn{\alpha} of the desired positive stable distribution (0 < a < 1).}
}
\value{
Numeric. Value of the constant \eqn{\gamma}.
}
\description{
Computes the normalization constant \eqn{\gamma} in the positive stable distribution
with index parameter \eqn{\alpha}.
}
\examples{
g(0.5)

}
