% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Data_Generation_TimeIndi_UTT.R
\name{gen_time_varying_cov_tx}
\alias{gen_time_varying_cov_tx}
\title{Generate Stratified Dataset with Time-Varying Covariates and Interval-Censored Survival Times}
\usage{
gen_time_varying_cov_tx(M, ni, beta, gen_visits, cens, rho = 0, baseline, S)
}
\arguments{
\item{M}{Numeric. Number of clusters or communities.}

\item{ni}{Numeric vector of length 2. Range (min, max) of individuals per community, sampled uniformly.}

\item{beta}{Numeric vector. Coefficients for: (1) the time-constant treatment variable, (2) the time-varying UTT indicator, and (3) their interaction.}

\item{gen_visits}{Function. Function to generate internal visit times (e.g., HIV testing dates).}

\item{cens}{Numeric. Loss-to-follow-up (censoring) rate. 0 implies no censoring.}

\item{rho}{Numeric. Kendall’s tau indicating within-cluster dependence. Set to 0 for independence.}

\item{baseline}{List. List of baseline distribution functions (generated from \code{return_baseline()} or \code{return_exp_baseline()}) for each stratum.}

\item{S}{Either a function that returns stratum assignment (e.g., sampling function), or a vector of length \code{M} giving fixed stratum assignments for each cluster.}
}
\value{
A data frame in long format with two rows per individual (before and after UTT policy) and the following columns:
\describe{
\item{id}{Individual ID.}
\item{stratum, group}{Stratum and cluster identifiers.}
\item{treat}{Cluster-assigned binary treatment indicator.}
\item{utt}{Time-varying UTT indicator (0 before policy, 1 after).}
\item{start, stop}{Start and stop times of the interval.}
\item{left, right}{Interval-censored survival times.}
}
}
\description{
Constructs a stratified, clustered dataset for a survival study with a binary treatment assignment
(constant within each cluster) and a time-varying intervention (UTT policy). The outcome is subject
to interval censoring and optional within-cluster dependence.
}
\details{
This function simulates data for survival models with interval censoring and time-varying covariates.
The time of UTT implementation varies across individuals. Dependence within clusters can be introduced via a
Clayton copula if \code{rho > 0}.
}
\examples{
set.seed(1)
baseline <- list(return_exp_baseline(10, 0.2, 0))
gen_visits <- function(max.time=10) sort(runif(5, 0, max.time))
simdat <- gen_time_varying_cov_tx(M = 4, ni = c(4, 6),
                                  beta = c(0.1, -0.2, 0.05),
                                  gen_visits = gen_visits,
                                  cens = 0.01,
                                  rho = 0,
                                  baseline = baseline,
                                  S = rep(1, 4))

}
