% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance_Estimation_Functions.R
\name{H}
\alias{H}
\title{Computes the Full-Data Sensitivity Matrix}
\usage{
H(beta, lambda.list, l, u, tau, x, mapping, p, h.n)
}
\arguments{
\item{beta}{Numeric vector of length \eqn{p}. Final fixed effect estimates.}

\item{lambda.list}{List. Profile maximum likelihood estimators for \code{lambda} under various first- and second-order perturbations of \code{beta}.}

\item{l}{List of length \eqn{S}. Each element is a numeric vector of left endpoints for each stratum.}

\item{u}{List of length \eqn{S}. Each element is a numeric vector of right endpoints for each stratum.}

\item{tau}{List of length \eqn{S}. Each element is a numeric vector of sorted unique timepoints for each stratum.}

\item{x}{List of length \eqn{S}. Each element is a covariate array for each stratum.}

\item{mapping}{List or \code{NULL}. Mapping of cluster indexing \eqn{(i,j)} to stratum indexing \eqn{(s,v)}; used for robust variance estimation. If \code{NULL}, clustering and stratification variables are assumed the same.}

\item{p}{Integer. Dimension of the coefficient vector \code{beta}.}

\item{h.n}{Numeric. Perturbation constant for finite difference approximation.}
}
\value{
A \eqn{p \times p} numeric matrix representing the full-data sensitivity (Hessian) matrix.
}
\description{
Calculates the sensitivity matrix (Hessian matrix) for the profile composite log-likelihood
with respect to the fixed effect estimates \code{beta} using finite difference approximations.
}
