% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance Estimation Functions: Profile
%   Likelihood.R
\name{J_i_1}
\alias{J_i_1}
\title{Cluster-Specific Variability Matrix Contribution (Same Stratification and Clustering)}
\usage{
J_i_1(beta, lambda.list.i, l.i, u.i, tau.i, x.i, p, h.n)
}
\arguments{
\item{beta}{Numeric vector of length \eqn{p}. Final fixed effect estimates.}

\item{lambda.list.i}{List. Contains the profile maximum likelihood estimators for the baseline hazard function of cluster \eqn{i} under:
\describe{
\item{\code{CMLE}}{Baseline hazard under the composite maximum likelihood estimator.}
\item{\code{first.order}}{List of length \eqn{p}, each containing the baseline hazard under a first-order perturbation of \code{beta}.}
}}

\item{l.i}{Numeric vector of length \eqn{n_i}. Left endpoints of the censoring intervals for individuals in cluster \eqn{i}.}

\item{u.i}{Numeric vector of length \eqn{n_i}. Right endpoints of the censoring intervals for individuals in cluster \eqn{i}.}

\item{tau.i}{Numeric vector of length \eqn{\rho_i}. Sorted unique timepoints for cluster (stratum) \eqn{i}.}

\item{x.i}{3D array. Covariate array for cluster \eqn{i}, where:
\itemize{
\item \code{x.i[j, , ]} returns a \eqn{p \times \rho_i} matrix for subject \eqn{j}.
\item \code{x.i[ , p, ]} returns an \eqn{n_i \times \rho_i} matrix for covariate \eqn{p}.
\item \code{x.i[ , , r]} returns an \eqn{n_i \times p} matrix for timepoint \eqn{r}.
}}

\item{p}{Integer. Dimension of the coefficient vector \code{beta}.}

\item{h.n}{Numeric. Perturbation constant used for finite difference approximation of the score.}
}
\value{
A \eqn{p \times p} matrix representing the outer product of the gradient (score vector) for cluster \eqn{i}.
}
\description{
Computes the contribution to the variability matrix from cluster \eqn{i}, assuming the stratification and clustering variables are the same.
}
\details{
This function contributes to the robust sandwich variance estimator by calculating the outer product of the score vector for a cluster under the assumption that clustering and stratification are the same. The score is approximated using finite differences.
}
\seealso{
\code{\link{composite_i_1}}, \code{\link{make_basis}}
}
