% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance_Estimation_Functions.R
\name{make_basis}
\alias{make_basis}
\title{Construct the k-th Canonical Basis Vector in \eqn{\mathbb{R}^p}}
\usage{
make_basis(k, p)
}
\arguments{
\item{k}{Integer. The index of the canonical basis vector to construct (1 ≤ k ≤ p).}

\item{p}{Integer. The dimension of the vector space.}
}
\value{
Numeric vector of length \eqn{p} with 1 at position \eqn{k} and 0 elsewhere.
}
\description{
Returns the \eqn{k}-th standard basis vector (canonical vector) of dimension \eqn{p},
which is a vector of zeros except a 1 in position \eqn{k}.
}
\examples{
make_basis(2, 5)
# [1] 0 1 0 0 0

}
