% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Processing Functions.R
\name{process_data}
\alias{process_data}
\title{Process Stratified Survival Data
Main wrapper that processes a dataset with optional stratification and clustering.}
\usage{
process_data(formula, data, strata, id, clus)
}
\arguments{
\item{formula}{formula object for the stratified COX model}

\item{data}{full set of observations}

\item{strata}{vector indicating the stratification variables}

\item{id}{individual subject identification variable}

\item{clus}{vector indicating the clustering variable}
}
\value{
A list with two components
\describe{
\item{pointEstimation}{A list containing processed data objects (like time intervals, risk sets, and covariates) ready for survival model estimation, organized by strata.}
\item{varMapping}{A list describing how clusters map to strata and individuals for variance estimation; or NULL if no clustering or stratification is used.}
}
}
\description{
Process Stratified Survival Data
Main wrapper that processes a dataset with optional stratification and clustering.
}
