% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maximization Step of the Composite EM
%   Algorithm.R
\name{update_lambda_s}
\alias{update_lambda_s}
\title{Update Baseline Hazard Components for a Stratum}
\usage{
update_lambda_s(beta, x.s, w.s, risk.vr, weights.s)
}
\arguments{
\item{beta}{Numeric vector of length \eqn{p}, representing the current estimates of the fixed-effect coefficients.}

\item{x.s}{3-dimensional array of covariates, where \code{x.s[v, , ]} returns a \eqn{p \times \rho_s} matrix for individual \eqn{v}; \code{x.s[, p, ]} gives an \eqn{n_s \times \rho_s} matrix for covariate \eqn{p}; and \code{x.s[, , r]} gives an \eqn{n_s \times p} matrix for timepoint \eqn{r}.}

\item{w.s}{Matrix (\eqn{n_s \times \rho_s}) of expected Poisson augmentation variables from the E-step, for stratum \eqn{s}.}

\item{risk.vr}{Matrix (\eqn{n_s \times \rho_s}) indicating risk set membership for each individual at each timepoint in stratum \eqn{s}.}

\item{weights.s}{Numeric vector of length \eqn{n_s}, containing perturbation weights for each individual in stratum \eqn{s}.}
}
\value{
A numeric vector of length \eqn{\rho_s}, containing the updated baseline hazard estimates \eqn{\lambda_{sr}} for stratum \eqn{s}.
}
\description{
Computes the updated baseline hazard estimates \eqn{\lambda_{sr}} at each timepoint \eqn{r} for a given stratum \eqn{s}, as part of the M-step in the composite EM algorithm.
}
\keyword{internal}
