% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maximization Step of the Composite EM
%   Algorithm.R
\name{x_bar_s}
\alias{x_bar_s}
\title{Computes the mean covariate vector at each timepoint in stratum s}
\usage{
x_bar_s(beta, u.s.star, tau.s, x.s, risk.vr, weights.s)
}
\arguments{
\item{beta}{Numeric vector of length \code{p}, the current fixed effect estimates.}

\item{u.s.star}{Numeric vector of length \code{n.s}, containing the right endpoints for each subject in stratum \code{s}.}

\item{tau.s}{Numeric vector of length \code{rho.s}, the sorted unique timepoints (left and right endpoints) for stratum \code{s}.}

\item{x.s}{Array of dimension \code{n.s x p x rho.s}, the covariate array for stratum \code{s}.}

\item{risk.vr}{Numeric matrix of dimension \code{n.s x rho.s}, indicating risk set membership for each subject at each timepoint.}

\item{weights.s}{Numeric vector of length \code{n.s}, containing perturbation weights for each individual in stratum \code{s}.}
}
\value{
A matrix of dimension \code{rho.s x p}, where each row gives the mean covariate vector at a specific timepoint.
}
\description{
This function calculates the weighted average of covariates at each timepoint \code{tau.s}
in a given stratum \code{s}, using the current estimate of the regression coefficients \code{beta}.
}
\keyword{internal}
